jobs_train/requirements.txt
datasets>=2.14.0
transformers>=4.30.0
accelerate>=0.26.0
torch>=2.0.0
google-cloud-storage>=2.0.0